Added handling for the /todos endpoint and updated the routing.
